Add session typings and introduce server session helper
This adds session types via types/next-auth.d.ts.

It also provides a helper function `auth()` that will automatically
redirect the user to the login page if there is no active session.
This way we know that the session object returned will never be null.
It also automatically provides the config so that we don't have to pass
it around.
