+++ src/app/admin/_actions/expense.ts
@@ -1,15 +1,15 @@
 'use server';
 
-import { authOptions } from '@/app/api/auth/[...nextauth]/route';
+import { auth } from '@/lib/auth';
 import { apiFetch, getFormValue } from '@/lib/helpers';
 import { apiUrl } from '@/lib/settings';
 import { NewExpense } from '@/lib/types';
-import { getServerSession } from 'next-auth';
 import { revalidatePath } from 'next/cache';
 
 export async function saveExpense(formData: FormData) {
   const responseBody = {} as NewExpense;
-  const session = await getServerSession(authOptions);
+  const session = await auth();
+  
   const method = formData.get('reqType') as string;
   formData.delete('reqType');
   formData.forEach((value, property: string) => {
@@ -21,7 +21,7 @@ export async function saveExpense(formData: FormData) {
 
   try {
     const res = await apiFetch(
-      (session as any)?.id_token,
+      session?.id_token,
       `${apiUrl}/adm/expenses`,
       {
         method,
@@ -46,12 +46,12 @@ export async function saveExpense(formData: FormData) {
   // redirect('/admin');
 }
 export async function deleteExpense(formData: FormData) {
-  const session = await getServerSession(authOptions);
+  const session = await auth();
   const id = formData.get('expenseId') as string;
 
   try {
     const res = await apiFetch(
-      (session as any)?.id_token,
+      session?.id_token,
       `${apiUrl}/adm/expenses?expenseId=${id}`,
       {
         method: 'DELETE',
