+++ src/app/mybudget/layout.tsx
@@ -1,22 +1,23 @@
 import Header from '@/components/Header';
-import { getServerSession } from 'next-auth';
-import { authOptions } from '../api/auth/[...nextauth]/route';
 import '../../styles/components/page.scss';
 import { redirect } from 'next/navigation';
+import { auth } from '@/lib/auth';
 
 export default async function Layout({
   children,
 }: {
   children: React.ReactNode;
 }) {
-  const session = await getServerSession(authOptions);
+  const session = await auth();
+
   if (!session) redirect('/api/auth/signin');
+
   return (
     <>
       <div className="page-wrapper">
         <Header
-          user={session?.user}
-          isAdmin={(session as any)?.isAdmin ?? false}
+          user={session.user}
+          isAdmin={session.isAdmin ?? false}
         />
         {children}
       </div>
